+++ doc/source/samples/optimal_res_scaling/green_PPA_sizing.py
@@ -7,13 +7,15 @@ from os.path import dirname, join
 import sys
 mypath = (dirname(__file__))
 sys.path.append(join(mypath, '../..'))
-import eao as eao
+sys.path.append(join(mypath, '../../..'))
+sys.path.append(join(mypath, '../../../..'))
+import eaopack as eao
 
-import eao
-from eao.assets import  Contract, Storage, SimpleContract, ScaledAsset
-from eao.basic_classes import Node, Timegrid, Unit
-from eao.portfolio import Portfolio
-from eao.serialization import load_from_json, run_from_json
+import eaopack as eao
+from eaopack.assets import  Contract, Storage, SimpleContract, ScaledAsset
+from eaopack.basic_classes import Node, Timegrid, Unit
+from eaopack.portfolio import Portfolio
+from eaopack.serialization import load_from_json, run_from_json
 
 ###############################################   setting
 ## input files
