+++ doc/source/samples/optimal_res_scaling/loop_charts.py
@@ -7,7 +7,8 @@ from os.path import dirname, join
 import sys
 mypath = (dirname(__file__))
 sys.path.append(join(mypath, '../../../..'))
-import eao as eao
+sys.path.append(join(mypath, '../../'))
+import eaopack as eao
 
 import matplotlib.pyplot as plt
