+++ doc/source/samples/optimal_res_scaling/vol_matched_vs_opt_charts.py
@@ -7,7 +7,7 @@ from os.path import dirname, join
 import sys
 mypath = (dirname(__file__))
 sys.path.append(join(mypath, '../../../..'))
-import eao as eao
+import eaopack as eao
 
 import matplotlib.pyplot as plt
