--- eao/__init__.py
@@ -1,8 +1,0 @@
-from eao import basic_classes
-from eao import assets
-from eao import portfolio
-from eao import optimization
-from eao import serialization
-from eao import network_graphs
-from eao import io
-from eao import stoch_lin_prog
