+++ eaopack/__init__.py
@@ -1,0 +1,8 @@
+from eaopack import basic_classes
+from eaopack import assets
+from eaopack import portfolio
+from eaopack import optimization
+from eaopack import serialization
+from eaopack import network_graphs
+from eaopack import io
+from eaopack import stoch_lin_prog
