+++ tests/test_optimization.py
@@ -9,7 +9,7 @@ from os.path import dirname, join
 import sys
 mypath = (dirname(__file__))
 sys.path.append(join(mypath, '..'))
-import eao as eao
+import eaopack as eao
 
 
 class OptimizationTests(unittest.TestCase):
